Added wall copy function
(cherry picked from commit 73ff9aab1f96390ca5c17cfb9e7c32ffd3072674)

+++ MapEditor/render/MapViewRenderer.cs
@@ -124,7 +124,7 @@ namespace MapEditor.render
         /// </summary>
         public void DrawTexturedWall(Graphics g, Map.Wall wall, bool transparent, bool beingSelected)
         {
-            ThingDb.Wall tt = ThingDb.Walls[wall.matId];
+            ThingDb.Wall tt = ThingDb.Walls[(int)wall.matId];
             try
             {
                 int actualVari = wall.Variation * 2;
